Updated split payment and spending reports.

+++ src/main/java/org/poo/bank/entity/ExchangeRate.java
@@ -3,10 +3,12 @@ package org.poo.bank.entity;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ExchangeRate {
     private String from;
     private String to;
-    private Double rate;
+    private BigDecimal rate;
 }
